updated some files

+++ Comidat.Data.Old/Data/Model/TBLMap.cs
@@ -6,44 +6,30 @@ namespace Comidat.Data.Model
 {
     public class TBLMap
     {
-        [Key]
-        [Obfuscation(Exclude = true)]
-        public int Id { get; set; }
+        [Key] [Obfuscation(Exclude = true)] public int Id { get; set; }
 
-        [Obfuscation(Exclude = true)]
-        public string MapName { get; set; }
+        [Obfuscation(Exclude = true)] public string MapName { get; set; }
 
-        [Obfuscation(Exclude = true)]
-        public string mp_image { get; set; }
+        [Obfuscation(Exclude = true)] public string mp_image { get; set; }
 
-        [Obfuscation(Exclude = true)]
-        public float MapWidth { get; set; }
+        [Obfuscation(Exclude = true)] public double MapWidth { get; set; }
 
-        [Obfuscation(Exclude = true)]
-        public float MapHeight { get; set; }
+        [Obfuscation(Exclude = true)] public double MapHeight { get; set; }
 
-        [Obfuscation(Exclude = true)]
-        public int Floor { get; set; }
+        [Obfuscation(Exclude = true)] public int Floor { get; set; }
 
-        [Obfuscation(Exclude = true)]
-        public double FloorDepth { get; set; }
+        [Obfuscation(Exclude = true)] public Single FloorDepth { get; set; }
 
-        [Obfuscation(Exclude = true)]
-        public bool MapActive { get; set; }
+        [Obfuscation(Exclude = true)] public bool MapActive { get; set; }
 
-        [Obfuscation(Exclude = true)]
-        public int RecordUserId { get; set; }
+        [Obfuscation(Exclude = true)] public int RecordUserId { get; set; }
 
-        [Obfuscation(Exclude = true)]
-        public DateTime RecordDateTime { get; set; }
+        [Obfuscation(Exclude = true)] public DateTime RecordDateTime { get; set; }
 
-        [Obfuscation(Exclude = true)]
-        public int? UpdateUserId { get; set; }
+        [Obfuscation(Exclude = true)] public int? UpdateUserId { get; set; }
 
-        [Obfuscation(Exclude = true)]
-        public DateTime? UpdateDateTime { get; set; }
+        [Obfuscation(Exclude = true)] public DateTime? UpdateDateTime { get; set; }
 
-        [Obfuscation(Exclude = true)]
-        public bool Deleted { get; set; }
+        [Obfuscation(Exclude = true)] public bool Deleted { get; set; }
     }
 }
